backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.13.3
pydantic-settings==2.14.0
python-dotenv==1.0.1
httpx==0.27.2

# LangChain core
langchain==0.3.1
langchain-core==0.3.6
langchain-community==0.3.1
langchain-text-splitters==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-chroma==0.1.4

# Vector store
chromadb==0.5.11

# Embeddings
sentence-transformers==3.1.1

# LLM provider
groq==0.11.0

# PDF export
reportlab==4.2.2

# Utilities
python-multipart==0.0.9
